Pull the remote item loader from the app container directly (#1)
The `remoteItemLoader` on the `Jigsaw` instance was marked as protected
at some point along the way. This made it impossible to access it
directly from the listener.

The same loader is defined in the app container and injected into the
Jigsaw instance though, and the `$jigsaw->app` variable is still
publicly accessible. So pulling the item loader from the accessible
container provides the same level of functionality. If the app property
were ever to be made protected or private this may no longer work,
though.